+++ pg/pgstorage/doc.go
@@ -1,2 +1,0 @@
-// Package pgstorage is reserved for PostgreSQL lazystorage metadata integration.
-package pgstorage